requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
authlib>=1.3.0
google-auth>=2.0.0
requests>=2.31.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Dev / test
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
